Add files via upload
XMODEM Transfer Protocol
Credit to professor Craig Scratchley for creating the outlines for the code, students were supposed to understand the outline extremely well to pass the tests as well as fill in the code required. As such a majority of this code is Craig Scratchley's while certain parts have been filled in. Certain documents that were given on that semester but were copyrighted aren't included in this repository, therefore this program will not run.